requirements.txt
flask_cors==3.0.9
Flask==2.1.2
Flask-DebugToolbar==0.11.0
gunicorn==20.0.4
fpdf
pdf2image
tesseract
bert-extractive-summarizer
sacremoses
pdfkit
nltk
yake
#tesseract-ocr 
pytesseract
transformers
